pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["James Winegar <jameswine@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
transformers = "^4.29.1"
torch = [
  {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_x86_64.whl", "platform" = "linux", markers = "platform_machine == 'x86_64'" },
  {url = "https://download.pytorch.org/whl/torch-2.0.1-cp310-cp310-manylinux2014_aarch64.whl", "platform" = "linux", markers = "platform_machine == 'aarch64'" },
  {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-win_amd64.whl", "platform" = "win32"},
  {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1-cp310-none-macosx_10_9_x86_64.whl", "platform" = "darwin",  markers = "platform_machine == 'x86_64'" },
  {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1-cp310-none-macosx_11_0_arm64.whl", "platform" = "darwin",  markers = "platform_machine == 'arm64'" }
]
scipy = "^1.11"
scikit-learn = "^1.3.0"
datasets = "^2.14"
requests = "^2.31"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}
fastapi = "^0.99"
uvicorn = "^0.23"

[tool.poetry.dev-dependencies]
pytest = "^7.4"
isort = "^5.12"
black = "^23.7"
pre-commit = "^3.3"

[tool.poetry.group.dev.dependencies]
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
